My 64-bit timer fix from September 28th wasn't enough. Attempting to calculate the tic was still being done in signed 32-bit integers elsewhere. Bit of a blitzkrieg to update that, but didn't get too far in the networking code because there's a ton of data type differences (and they won't get "fixed" since that breaks vanilla compatibility so just need a ton of casts instead).
Left a debugging note in i_timer for anyone else wanting to play with it.

+++ src/doom/g_game.h
@@ -64,7 +64,7 @@ void G_WorldDone (void);
 
 // Read current data from inputs and build a player movement command.
 
-void G_BuildTiccmd (ticcmd_t *cmd, int maketic); 
+void G_BuildTiccmd (ticcmd_t *cmd, uint64_t maketic); 
 
 void G_Ticker (void);
 boolean G_Responder (event_t*	ev);
